+++ src/components/ContactForm.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import toast from 'react-hot-toast'
 import Link from '~/components/Link'
 import type { FormEvent } from 'react'
